refactor binner and slicer to reuse code
